python/requirements.txt
# Core data + math
pandas
numpy

# Market data (no API key needed)
yfinance

# Load environment variables from .env
python-dotenv

# Testing
pytest

# Useful for HTTP later (stretch / optional but good)
requests